+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Add colors to example HTML
 - Update example configurations to use Team Widget
 - Add Team Widget support to Companion module
+- Fixed but in reset counter using first increment value instead of initial\_value
+- Expand widget documentation and move to github wiki.
 
 ### Bug Fixes